# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.JSONField:
    additionalProperties: true
    type: object
...
  entity.MessageQueue:
    properties:
      completed_at:
        type: string
      created_at:
        example: "2023-01-01T12:00:00Z"
        type: string
      deleted_at:
        example: "2023-01-01T12:00:00Z"
        format: date-time
        type: string
      error_message:
        type: string
      id:
        example: 1
        type: integer
      max_retries:
        type: integer
      message_id:
        type: string
      payload:
        $ref: '#/definitions/entity.JSONField'
      priority:
        type: integer
      queue_name:
        type: string
      retry_count:
        type: integer
      scheduled_at:
        type: string
      started_at:
        type: string
      status:
        $ref: '#/definitions/entity.QueueStatus'
      updated_at:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
...
  entity.QueueStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-comments:
      QueueStatusCancelled: 已取消
      QueueStatusCompleted: 已完成
      QueueStatusFailed: 处理失败
      QueueStatusPending: 待处理
      QueueStatusRetrying: 重试中
      QueueStatusRunning: 运行中
    x-enum-varnames:
    - QueueStatusPending
    - QueueStatusRunning
    - QueueStatusCompleted
    - QueueStatusFailed
    - QueueStatusRetrying
    - QueueStatusCancelled
...
  usecase.QueueMessageListResult:
    properties:
      items:
        items:
          $ref: '#/definitions/entity.MessageQueue'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  usecase.QueueStatsResult:
    properties:
      completed_jobs:
        type: integer
      failed_jobs:
        type: integer
      pending_jobs:
        type: integer
      running_jobs:
        type: integer
      total_jobs:
        type: integer
    type: object
...
  v1.StandardResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 统一消息机器人管理平台API
  title: OmniBotGo API
  version: "1.0"
paths:
...
  /api/v1/queues:
    get:
      consumes:
      - application/json
      description: 获取消息队列中的消息列表，支持分页和过滤
      parameters:
      - description: 消息状态
        in: query
        name: status
        type: string
      - description: 队列类型
        in: query
        name: queue_type
        type: string
      - description: 优先级
        in: query
        name: priority
        type: integer
      - default: 1
        description: 页码，默认1
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认20
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.QueueMessageListResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取队列消息列表
      tags:
      - queues
  /api/v1/queues/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取指定队列消息的详细信息
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/entity.MessageQueue'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 消息不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取队列消息详情
      tags:
      - queues
  /api/v1/queues/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: 取消指定的队列消息处理
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取消成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 消息不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 取消队列消息
      tags:
      - queues
  /api/v1/queues/{id}/retry:
    post:
      consumes:
      - application/json
      description: 重新处理指定的队列消息
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 重试成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "404":
          description: 消息不存在
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 重试队列消息
      tags:
      - queues
  /api/v1/queues/failed:
    delete:
      consumes:
      - application/json
      description: 清理队列中的失败消息
      parameters:
      - description: 清理指定时间之前的失败消息
        in: query
        name: before
        type: string
      - default: 100
        description: 批量处理大小，默认100
        in: query
        name: batch_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 清理成功
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 清理失败消息
      tags:
      - queues
  /api/v1/queues/stats:
    get:
      consumes:
      - application/json
      description: 获取消息队列的统计信息，包括处理情况、性能指标等
      parameters:
      - description: 队列类型
        in: query
        name: queue_type
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/v1.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/usecase.QueueStatsResult'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/v1.StandardResponse'
      summary: 获取队列统计信息
      tags:
      - queues
...
swagger: "2.0"